server: mob variants in all their vanilla coats
Roll, inherit, persist and sync the visual variants of wolf (coat by
biome tag, one per pack), cat (the eleven coats; all-black under a full
moon and in the swamp hut's tag), horse (colour per herd, markings per
animal, foals blending their parents' odds), llama and trader llama
(coat and strength, shared by the group), parrot, rabbit (biome types,
one in twenty a fresh roll on breeding), fox (snow or red by biome),
mooshroom (red) and the temperature variants of pig, cow and chicken
(warm/cold/temperate by biome tag) — each from the vanilla spawn and
breeding rule, carried through the vanilla spawn-group data so a pack
shares one look. The metadata rides on the synced index and serializer
each species uses on 1.21.5 and 26.2 (tachyne-common 73d9aa5 carries
the per-version renumbering, and fixes the villager data serializer to
19). Persisted variants widen to int32; rows without one re-roll on
load; a saved llama strength only overrides a fresh roll when set.

+++ internal/server/mob.go
@@ -77,7 +77,7 @@ type mob struct {
 	customName      string   // name-tagged: shown above the mob, and it never despawns
 	fromBucket      bool     // released from a mob bucket: persistent (Bucketable.setFromBucket)
 	persistent      bool     // Mob.persistenceRequired: picked up gear (never despawns)
-	variant         int8     // frog / axolotl colour (variant.go); meaningful when variantSet
+	variant         int32    // species variant (variant.go: coat/colour, horse colour|markings<<8, villager type); meaningful when variantSet
 	variantSet      bool
 	eggIn           int        // chicken: ticks until the next egg
 	beeNectar       bool       // bee: carrying nectar home (fills the hive on delivery)
